+++ grafast/dataplan-pg/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @dataplan/pg
 
+## 1.0.3
+
+### Patch Changes
+
+- [#3051](https://github.com/graphile/crystal/pull/3051)
+  [`e7f9242`](https://github.com/graphile/crystal/commit/e7f9242a5f592c7dca590e44457b500e1f9f5178)
+  Thanks [@benjie](https://github.com/benjie)! - Add new `RECORD_EXPRESSION`
+  mode to PgCondition so postgraphile-plugin-connection-filter doesn't need to
+  rely on a hack to enable filtering on composite columns. Also changes
+  `PgCondition` signature to accept a configuration object.
+
+- [#3050](https://github.com/graphile/crystal/pull/3050)
+  [`75a4c69`](https://github.com/graphile/crystal/commit/75a4c692aa90db63e5442cd523b5b94acc7afef0)
+  Thanks [@benjie](https://github.com/benjie)! - Add new
+  `PgCondition::ignoreUnlessAmended()` method to have PgCondition modifiers only
+  apply if additional conditions are added. Needed to fix a regression in
+  postgraphile-plugin-connection-filter
+- Updated dependencies
+  [[`0c7ec9d`](https://github.com/graphile/crystal/commit/0c7ec9d512c847b96952fd6864babbe8be2f1d3b),
+  [`ac376ca`](https://github.com/graphile/crystal/commit/ac376ca8d5fa76c0654b9256e049519bfd34172d),
+  [`9600dfa`](https://github.com/graphile/crystal/commit/9600dfa0aa67a215695ad522df09580dac061de8)]:
+  - grafast@1.0.2
+
+## 1.0.2
+
+### Patch Changes
+
+- [#3034](https://github.com/graphile/crystal/pull/3034)
+  [`582476e`](https://github.com/graphile/crystal/commit/582476e69c58a6e192dec8be53adf590dfa9c769)
+  Thanks [@benjaie](https://github.com/benjaie)! - Add `TYPES.unknown` codec to
+  make porting from V4 easier
+
 ## 1.0.1
 
 ### Patch Changes